Split kaapi_tasklist_t in two parts:
- kaapi_frame_tasklist_t is the other representation of the tasks into the frame
Once build, this data structure (should) never change.
- kaapi_tasklist_t is the executive part of the old kaapi_tasklist_t.

Solve most of the #include problem.
[for JOAO]: only kaapi_impl.h and few header files is assumed to be directly
included in the source files. There are a lot of dependency between includes which
are solve in kaapi_impl.h....


Change the thread_execframe to compile. Should be tests.

DO NOT INCLUDE THIS COMMIT YET - NOT TESTED

Conflicts resolved:
	src/machine/mt/kaapi_mt_sched_emitsteal.c
	src/workstealing/kaapi_sched_stealprocessor.c

+++ src/machine/mt/kaapi_mt_machine.h
@@ -70,9 +70,12 @@
 #if defined(KAAPI_USE_CUDA)
 #  include "../cuda/kaapi_cuda_proc.h"
 #endif
-#include "../../memory/kaapi_mem.h"
 
 #include "kaapi_mt_bitmap.h"
+
+#include "../../memory/kaapi_mem.h"
+#include "../../memory/kaapi_mem_data.h"
+#include "../../memory/kaapi_mem_host_map.h"
 
 /* ========================================================================== */
 struct kaapi_procinfo_t;
@@ -254,11 +257,11 @@ static inline void kaapi_listrequest_iterator_unset_at(
 
 /* return the number of entries in the range
 */
-static inline unsigned long kaapi_listrequest_iterator_count(
+static inline int kaapi_listrequest_iterator_count(
     kaapi_listrequest_iterator_t* lrrange
 )
 { return kaapi_bitmap_value_count(&lrrange->bitmap) 
-      + (unsigned long)(lrrange->idcurr == -1 ? 0 : 1); }
+      + (int)(lrrange->idcurr == -1 ? 0 : 1); }
 
 /* get the first request of the range. range iterator should have been initialized by kaapi_listrequest_iterator_init 
 */
